+++ visualization/app/codeCharta/state/effects/nodeContextMenu/excludeButton/excludeButton.component.ts
@@ -1,0 +1,26 @@
+import { Component, Inject, Input } from "@angular/core"
+import { BlacklistType, CodeMapNode } from "../../../../codeCharta.model"
+import { Store } from "../../../angular-redux/store"
+import { addBlacklistItemsIfNotResultsInEmptyMap } from "../../../store/fileSettings/blacklist/blacklist.actions"
+
+@Component({
+	selector: "cc-exclude-button",
+	template: require("./excludeButton.component.html")
+})
+export class ExcludeButtonComponent {
+	@Input() codeMapNode: Pick<CodeMapNode, "path" | "type">
+
+	constructor(@Inject(Store) private store: Store) {}
+
+	excludeNode() {
+		this.store.dispatch(
+			addBlacklistItemsIfNotResultsInEmptyMap([
+				{
+					path: this.codeMapNode.path,
+					type: BlacklistType.exclude,
+					nodeType: this.codeMapNode.type
+				}
+			])
+		)
+	}
+}