+++ visualization/app/codeCharta/util/usageDataTracker.ts
@@ -1,7 +1,7 @@
 "use strict"
 
 import { getVisibleFileStates, isSingleState } from "../model/files/files.helper"
-import { CodeMapNode, NodeType, State } from "../codeCharta.model"
+import { CodeMapNode, NodeType } from "../codeCharta.model"
 import { isActionOfType } from "./reduxHelper"
 import { AreaMetricActions } from "../state/store/dynamicSettings/areaMetric/areaMetric.actions"
 import { HeightMetricActions } from "../state/store/dynamicSettings/heightMetric/heightMetric.actions"
@@ -15,6 +15,7 @@ import { hierarchy } from "d3-hierarchy"
 import { getMedian, pushSorted } from "./nodeDecorator"
 import { RangeSliderController } from "../ui/rangeSlider/rangeSlider.component"
 import { ColorRangeActions } from "../state/store/dynamicSettings/colorRange/colorRange.actions"
+import { FileState } from "../model/files/files"
 
 interface MetaDataTrackingItem {
 	mapId: string
@@ -56,10 +57,10 @@ interface MetricStatistics {
 
 export const TRACKING_DATA_LOCAL_STORAGE_ELEMENT = "CodeCharta::usageData"
 
-function isTrackingAllowed(state: State) {
-	const singleFileStates = getVisibleFileStates(state.files)
+function isTrackingAllowed(files: FileState[]) {
+	const singleFileStates = getVisibleFileStates(files)
 
-	if (!isSingleState(state.files) || singleFileStates.length > 1) {
+	if (!isSingleState(files) || singleFileStates.length > 1) {
 		return false
 	}
 
@@ -72,12 +73,12 @@ function isTrackingAllowed(state: State) {
 	)
 }
 
-export function trackMapMetaData(state: State) {
-	if (!isTrackingAllowed(state)) {
+export function trackMapMetaData(files: FileState[]) {
+	if (!isTrackingAllowed(files)) {
 		return
 	}
 
-	const singleFileStates = getVisibleFileStates(state.files)
+	const singleFileStates = getVisibleFileStates(files)
 	const fileNodes: CodeMapNode[] = getFileNodes(singleFileStates[0].file.map)
 	const fileMeta = singleFileStates[0].file.fileMeta
 
@@ -268,9 +269,9 @@ interface EventTrackingItem {
 	payload: SettingChangedEventPayload | NodeInteractionEventPayload
 }
 
-export function trackEventUsageData(actionType: string, state: State, payload?: any) {
+export function trackEventUsageData(actionType: string, files: FileState[], payload?: any) {
 	if (
-		!isTrackingAllowed(state) ||
+		!isTrackingAllowed(files) ||
 		(!isActionOfType(actionType, AreaMetricActions) &&
 			!isActionOfType(actionType, HeightMetricActions) &&
 			!isActionOfType(actionType, ColorMetricActions) &&
@@ -282,7 +283,7 @@ export function trackEventUsageData(actionType: string, state: State, payload?: 
 		return
 	}
 
-	const singleFileStates = getVisibleFileStates(state.files)
+	const singleFileStates = getVisibleFileStates(files)
 	const fileMeta = singleFileStates[0].file.fileMeta
 
 	const eventTrackingItem = buildEventTrackingItem(fileMeta.fileChecksum, actionType, payload)
@@ -291,7 +292,7 @@ export function trackEventUsageData(actionType: string, state: State, payload?: 
 	}
 
 	// Make sure that only files within usageData can be read
-	const fileChecksum = getVisibleFileStates(state.files)[0].file.fileMeta.fileChecksum.replace(/\//g, "")
+	const fileChecksum = getVisibleFileStates(files)[0].file.fileMeta.fileChecksum.replace(/\//g, "")
 
 	let appendedEvents = ""
 	try {