+++ visualization/app/material/material.module.ts
@@ -8,6 +8,8 @@ import { MatTooltipModule } from "@angular/material/tooltip"
 import { MatIconModule } from "@angular/material/icon"
 import { MatButtonToggleModule } from "@angular/material/button-toggle"
 import { MatDialogModule } from "@angular/material/dialog"
+import { MatCardModule } from "@angular/material/card"
+import { MatCheckboxModule } from "@angular/material/checkbox"
 
 const materialModules = [
 	MatSelectModule,
@@ -16,8 +18,11 @@ const materialModules = [
 	MatDividerModule,
 	MatTooltipModule,
 	MatIconModule,
+	MatDialogModule,
+	MatCardModule,
 	MatButtonToggleModule,
-	MatDialogModule
+	MatDialogModule,
+	MatCheckboxModule
 ]
 
 @NgModule({